src/test/java/org/vortikal/text/html/HtmlDigesterTest.java
/* Copyright (c) 2012, University of Oslo, Norway
 * All rights reserved.
 * 
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions are
 * met:
 * 
 *  * Redistributions of source code must retain the above copyright
 *    notice, this list of conditions and the following disclaimer.
 * 
 *  * Redistributions in binary form must reproduce the above copyright
 *    notice, this list of conditions and the following disclaimer in the
 *    documentation and/or other materials provided with the distribution.
 * 
 *  * Neither the name of the University of Oslo nor the names of its
 *    contributors may be used to endorse or promote products derived from
 *    this software without specific prior written permission.
 *      
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
 * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
 * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
 * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
 * OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
 * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
 * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
 * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
 * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
 * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
package org.vortikal.text.html;

import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;

import org.junit.Before;
import org.junit.Test;

public class HtmlDigesterTest {

    private HtmlDigester htmlDigester;
    private String simpleHtmlParagraph;

    @Before
    public void init() {
        this.htmlDigester = new HtmlDigester();
        this.simpleHtmlParagraph = "<p>Some bold <b>text</b> and <i>italic</i>. A list:</p>"
                + "<ul><li>element 1</li><li>element 2</li><li>element 3</li></ul>";
    }

    @Test
    public void truncateHtmlAlreadyWithinLimit() {
        this.truncateHtml(this.simpleHtmlParagraph, 500);
    }

    @Test
    public void truncateSimpleHtml() {
        this.truncateHtml(this.simpleHtmlParagraph, 50);
    }

    private void truncateHtml(String html, int limit) {
        this.htmlDigester.setLimit(limit);
        String truncated = this.htmlDigester.truncateHtml(html, limit);
        assertNotNull(truncated);
        assertTrue(truncated.length() <= limit);
    }

}
